CPM-1016: Add Structure Context
