+++ components/identifier-generator/back/tests/Acceptance/Context/Condition/BaseCreateOrUpdateConditionContext.php
@@ -30,33 +30,7 @@ class BaseCreateOrUpdateConditionContext extends BaseCreateOrUpdateIdentifierGen
         string $value = '',
         string $unknown = '',
     ): void {
-        $defaultCondition = $this->getValidCondition($type);
-        if ($attributeCode !== '') {
-            $defaultCondition['attributeCode'] = $attributeCode;
-        }
-        if ('undefined' === $scope) {
-            unset($defaultCondition['scope']);
-        } elseif ('' !== $scope) {
-            $defaultCondition['scope'] = $scope;
-        }
-        if ('undefined' === $locale) {
-            unset($defaultCondition['locale']);
-        } elseif ('' !== $locale) {
-            $defaultCondition['locale'] = $locale;
-        }
-        if ('undefined' === $value) {
-            unset($defaultCondition['value']);
-        } elseif ($value !== '') {
-            $defaultCondition['value'] = \json_decode($value);
-        }
-        if ('undefined' === $operator) {
-            unset($defaultCondition['operator']);
-        } elseif ($operator !== '') {
-            $defaultCondition['operator'] = $operator;
-        }
-        if ($unknown !== '') {
-            $defaultCondition['unknown'] = 'unknown property';
-        }
+        $defaultCondition = $this->getCondition($type, $attributeCode, $scope, $locale, $value, $operator, $unknown);
         $this->tryToCreateGenerator(conditions: [$defaultCondition]);
     }
 
@@ -79,6 +53,19 @@ class BaseCreateOrUpdateConditionContext extends BaseCreateOrUpdateIdentifierGen
         string $value = '',
         string $unknown = '',
     ): void {
+        $defaultCondition = $this->getCondition($type, $attributeCode, $scope, $locale, $value, $operator, $unknown);
+        $this->tryToUpdateGenerator(conditions: [$defaultCondition]);
+    }
+
+    private function getCondition(
+        string $type,
+        string $attributeCode,
+        string $scope,
+        string $locale,
+        string $value,
+        string $operator,
+        string $unknown
+    ): array {
         $defaultCondition = $this->getValidCondition($type);
         if ($attributeCode !== '') {
             $defaultCondition['attributeCode'] = $attributeCode;
@@ -106,35 +93,7 @@ class BaseCreateOrUpdateConditionContext extends BaseCreateOrUpdateIdentifierGen
         if ($unknown !== '') {
             $defaultCondition['unknown'] = 'unknown property';
         }
-        $this->tryToUpdateGenerator(conditions: [$defaultCondition]);
-    }
 
-    private function getValidCondition(string $type, ?string $operator = null): array
-    {
-        switch($type) {
-            case 'enabled': return [
-                'type' => 'enabled',
-                'value' => true,
-            ];
-            case 'family': return [
-                'type' => 'family',
-                'operator' => $operator ?? 'IN',
-                'value' => ['tshirt'],
-            ];
-            case 'simple_select': return [
-                'type' => 'simple_select',
-                'operator' => $operator ?? 'IN',
-                'attributeCode' => 'color',
-                'value' => ['green'],
-            ];
-            case 'multi_select': return [
-                'type' => 'multi_select',
-                'operator' => $operator ?? 'IN',
-                'attributeCode' => 'a_multi_select',
-                'value' => ['option_a', 'option_b'],
-            ];
-        }
-
-        throw new \InvalidArgumentException('Unknown type ' . $type . ' for getValidCondition');
+        return $defaultCondition;
     }
 }
