+++ components/identifier-generator/back/tests/features/createFamilyNomenclature.feature
@@ -3,6 +3,6 @@ Feature: Create Family Nomenclature
 
   Scenario: Can create the nomenclature operator
     When I create the family nomenclature operator to <=, value to 5 and generation if empty
-    Then I should not have an error
+    Then I should not get any error
     Then the family nomenclature operator should be <=
     And the family nomenclature value should be 5
