+++ components/identifier-generator/back/tests/features/createSimpleSelectNomenclature.feature
@@ -6,6 +6,6 @@ Feature: Create Simple Select Nomenclature
 
   Scenario: Can create the simple select nomenclature
     When I create the simple select nomenclature of attribute color operator to <=, value to 5 and generation if empty
-    Then I should not have an error
+    Then I should not get any error
     Then the simple select nomenclature operator for color should be <=
     And the color simple select nomenclature value should be 5
